add tests, fix export to csv
